Split appShared.ts into focussed modules
