+++ frontend/glovelly-web/src/components/AdminSection.tsx
@@ -1,6 +1,6 @@
 import type { FormEvent } from 'react'
-import type { AdminUser, AdminUserForm } from '../appShared'
-import { formatDateTime } from '../appShared'
+import { formatDateTime } from '../formatters'
+import type { AdminUser, AdminUserForm } from '../types'
 
 type AdminSectionProps = {
   adminForm: AdminUserForm
