+++ frontend/glovelly-web/src/components/InvoicesSection.tsx
@@ -1,5 +1,10 @@
-import type { Invoice, InvoiceStatus } from '../appShared'
-import { formatCurrency, formatDate, formatDateTime, getAllowedInvoiceStatusTransitions } from '../appShared'
+import {
+  formatCurrency,
+  formatDate,
+  formatDateTime,
+  getAllowedInvoiceStatusTransitions,
+} from '../formatters'
+import type { Invoice, InvoiceStatus } from '../types'
 
 type InvoicesSectionProps = {
   adjustmentAmount: string
