+++ frontend/glovelly-web/src/components/SessionCheckingScreen.tsx
@@ -1,5 +1,5 @@
-import type { AppMetadata } from '../appShared'
-import { formatBuildMetadata } from '../appShared'
+import { formatBuildMetadata } from '../formatters'
+import type { AppMetadata } from '../types'
 
 type SessionCheckingScreenProps = {
   appMetadata: AppMetadata
