+++ frontend/glovelly-web/src/theme.ts
@@ -1,0 +1,16 @@
+import type { ThemePreference } from './types'
+
+export const themeStorageKey = 'glovelly.theme-preference'
+
+export function getStoredThemePreference(): ThemePreference {
+  const rawPreference = window.localStorage.getItem(themeStorageKey)
+  if (
+    rawPreference === 'system' ||
+    rawPreference === 'light' ||
+    rawPreference === 'dark'
+  ) {
+    return rawPreference
+  }
+
+  return 'system'
+}
